level 31 - 35 & changes
levels 31 - 35 completed.

made a few changes to rubber to act differently. Will now be allowed to
be hit by dynamite again after it has come to a stop. Also it will move
5 blocks horizontally before descending.

GAME DELAYED!!!!!
5Grindr apologises for delays to it's fans and promises that it will be
completed by Saturday 7th March and will include a bonus in the hopes to
appease the loyal fanbase.